Relocate meta stuff
